# infrastructure/src/main/webui/src/api/model/matchedOneToOneLifecycle.ts
/**
 * Generated by orval v6.24.0 🍺
 * Do not edit manually.
 * infrastructure API
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 */
import type { MatchedStatus } from "./matchedStatus";

export interface MatchedOneToOneLifecycle {
  initiatedBy: string;
  joinedBy: string;
  leavedAt?: string;
  matchedAt: string;
  matchedStatus: MatchedStatus;
}