# infrastructure/src/main/webui/src/api/model/matchedStatus.ts
/**
 * Generated by orval v6.24.0 🍺
 * Do not edit manually.
 * infrastructure API
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 */

export type MatchedStatus = (typeof MatchedStatus)[keyof typeof MatchedStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MatchedStatus = {
  MATCHED: "MATCHED",
  LEAVED_INITIATOR: "LEAVED_INITIATOR",
  LEAVED_STRANGER: "LEAVED_STRANGER",
} as const;