# infrastructure/src/main/webui/src/api/model/mediaType.ts
/**
 * Generated by orval v6.24.0 🍺
 * Do not edit manually.
 * infrastructure API
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 */
import type { MediaTypeParameters } from "./mediaTypeParameters";

export interface MediaType {
  hash?: number;
  parameters?: MediaTypeParameters;
  subtype?: string;
  type?: string;
  wildcardSubtype?: boolean;
  wildcardType?: boolean;
}