# infrastructure/src/main/webui/src/api/model/outboundSseEvent.ts
/**
 * Generated by orval v6.24.0 🍺
 * Do not edit manually.
 * infrastructure API
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 */
import type { OutboundSseEventGenericType } from "./outboundSseEventGenericType";
import type { MediaType } from "./mediaType";
import type { OutboundSseEventType } from "./outboundSseEventType";

export interface OutboundSseEvent {
  comment?: string;
  data?: unknown;
  genericType?: OutboundSseEventGenericType;
  id?: string;
  mediaType?: MediaType;
  name?: string;
  reconnectDelay?: number;
  reconnectDelaySet?: boolean;
  type?: OutboundSseEventType;
}